New implementation of the poset - changes in header mcem: - Replace Eigen::MatrixXi with BGL adjacency list
